Prevent user that already answered consultation to answer again

+++ src/main/kotlin/fr/social/gouv/agora/usecase/reponseConsultation/InsertReponseConsultationUseCase.kt
@@ -2,16 +2,17 @@ package fr.social.gouv.agora.usecase.reponseConsultation
 
 import fr.social.gouv.agora.domain.ReponseConsultationInserting
 import fr.social.gouv.agora.usecase.consultation.repository.ConsultationPreviewAnsweredRepository
+import fr.social.gouv.agora.usecase.reponseConsultation.repository.GetConsultationResponseRepository
 import fr.social.gouv.agora.usecase.reponseConsultation.repository.InsertReponseConsultationRepository
-import fr.social.gouv.agora.usecase.reponseConsultation.repository.InsertReponseConsultationRepository.InsertParameters
 import fr.social.gouv.agora.usecase.reponseConsultation.repository.InsertReponseConsultationRepository.InsertResult
 import org.springframework.stereotype.Service
-import java.util.*
 
 @Service
 class InsertReponseConsultationUseCase(
     private val consultationPreviewAnsweredRepository: ConsultationPreviewAnsweredRepository,
     private val insertReponseConsultationRepository: InsertReponseConsultationRepository,
+    private val consultationResponseRepository: GetConsultationResponseRepository,
+    private val insertConsultationResponseParametersMapper: InsertConsultationResponseParametersMapper,
 ) {
 
     fun insertReponseConsultation(
@@ -19,12 +20,15 @@ class InsertReponseConsultationUseCase(
         userId: String,
         consultationResponses: List<ReponseConsultationInserting>,
     ): InsertResult {
+        if (consultationResponseRepository.hasAnsweredConsultation(consultationId = consultationId, userId = userId)) {
+            return InsertResult.INSERT_FAILURE
+        }
+
         consultationPreviewAnsweredRepository.deleteConsultationAnsweredList(userId)
         return insertReponseConsultationRepository.insertConsultationResponses(
-            insertParameters = InsertParameters(
+            insertParameters = insertConsultationResponseParametersMapper.toInsertParameters(
                 consultationId = consultationId,
                 userId = userId,
-                participationId = UUID.randomUUID().toString(),
             ),
             consultationResponses = consultationResponses,
         )
